requirements.agent.txt
# Agent Requirements (NO TENSORFLOW)
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Networking and messaging
redis>=4.0.0
kafka-python>=2.0.0
requests>=2.25.0

# Data processing
pydantic>=1.8.0
python-dateutil>=2.8.0

# Logging and monitoring
structlog>=21.1.0
prometheus-client>=0.11.0

# Utilities
click>=8.0.0
python-dotenv>=0.19.0
psutil>=5.8.0

# Explicitly exclude TensorFlow to prevent mutex issues
# tensorflow is NOT included here
